+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## v1.0.0-alpha.4
 
+### Focus
+
+Implement features requested by the client/custormer.
+
+### Changes
+
+- Implemented a large departure time countdown at the bottom of the screen.
 
 ## v1.0.0-alpha.3